refactor: remove unused CRUD helper and related interfaces; update transaction DTOs and GraphQL schema for improved transaction handling
