--- src/graphql/dtos/transaction.dto.ts
@@ -1,56 +1,0 @@
-import { ObjectType, Field } from '@nestjs/graphql';
-import { IsEthereumAddress, IsOptional } from 'class-validator';
-
-@ObjectType()
-export class TransactionDto {
-  @Field()
-  hash: string;
-
-  @Field()
-  blockHash: string;
-
-  @Field()
-  nonce: string;
-
-  @Field()
-  @IsEthereumAddress()
-  from: string;
-
-  @Field()
-  @IsEthereumAddress()
-  @IsOptional()
-  to: string;
-
-  @Field()
-  value: string;
-
-  @Field()
-  gas: string;
-
-  @Field({ nullable: true })
-  timestamp: Date;
-
-  @Field()
-  blockNumber: string;
-
-  @Field({ nullable: true })
-  asset?: string;
-
-  @Field({ nullable: true })
-  isError: string;
-
-  @Field({ nullable: true })
-  txreceipt_status?: string;
-
-  @Field({ nullable: true })
-  contractAddress: string;
-
-  @Field({ nullable: true })
-  cumulativeGasUsed?: string;
-
-  @Field({ nullable: true })
-  gasUsed?: string;
-
-  @Field({ nullable: true })
-  confirmations?: number;
-}
